Create clean test for inline Numeric
Split code into separate files for clarity

Reimplement Numeric and Ordering as inline

Place symbols in the trait
